Move out LightTypeTagImpl instead
