--- fundamentals/fundamentals-reflection/src/main/scala/izumi/fundamentals/reflection/StableType.scala
@@ -1,3 +1,0 @@
-package izumi.fundamentals.reflection
-
-import izumi.fundamentals.reflection.ReflectionUtil._
